third_party/WebKit/LayoutTests/svg/dynamic-updates/SVGLineElement-dom-requiredFeatures-expected.txt

SVG 1.1 dynamic update tests

Check that SVGLineElement is initially displayed
PASS lineElement.getBBox().width is 180
PASS lineElement.getBBox().height is 180
Check that setting requiredFeatures to something invalid makes it not render
PASS lineElement.getBBox().width is 0
PASS lineElement.getBBox().height is 0
Check that setting requiredFeatures to something valid makes it render again
PASS lineElement.getBBox().width is 180
PASS lineElement.getBBox().height is 180
Check that adding something valid to requiredFeatures keeps rendering the element
PASS lineElement.getBBox().width is 180
PASS lineElement.getBBox().height is 180
Check that adding something invalid to requiredFeatures makes it not render
PASS lineElement.getBBox().width is 0
PASS lineElement.getBBox().height is 0
PASS successfullyParsed is true

TEST COMPLETE
